feat: improve quiz UX and active quiz flow - Update docs and add new questions - Support fractional spaced-rep intervals (logic + DTOs) - Prefill JoinGameForm and add QuizBrowser - Add active quizzes API (controller/service + Firestore) - Include per-question results in quiz-completed webhook

+++ spring-backend/src/main/java/com/myproject/quizzai/dto/ReviewScheduleResponseDto.java
@@ -17,7 +17,7 @@ public class ReviewScheduleResponseDto {
     private String user_id;
     private String category;
     private double easiness;
-    private int interval_days;
+    private double interval_days;
     private int repetitions;
     private String next_review;
     private String last_reviewed;
